Refactor: 🔨 결제 테스트 패키지 리팩토링
기존 테스트의 경로설정을 수정하는 바람에 실패가 뜨는 테스트 코드의 경로를 바로잡았습니다.
