+++ src/test/java/profect/eatcloud/Domain/Payment/Service/PointServiceTest.java
@@ -1,4 +1,4 @@
-package profect.eatcloud.Domain.Payment;
+package profect.eatcloud.Domain.Payment.Service;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
@@ -15,7 +15,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import profect.eatcloud.Domain.Customer.Entity.Customer;
 import profect.eatcloud.Domain.Customer.Repository.CustomerRepository;
-import profect.eatcloud.Domain.Payment.Service.PointService;
 import profect.eatcloud.Domain.Payment.Service.PointService.PointResult;
 
 @ExtendWith(MockitoExtension.class)
@@ -30,7 +29,6 @@ class PointServiceTest {
     @DisplayName("포인트 충분 시 사용 성공")
     @Test
     void givenSufficientPoints_whenUsePoints_thenReturnSuccess() {
-        // given
         UUID customerId = UUID.randomUUID();
         Customer customer = Customer.builder()
                 .id(customerId)
@@ -42,10 +40,8 @@ class PointServiceTest {
         given(customerRepository.save(any(Customer.class)))
                 .willReturn(customer);
 
-        // when
         PointResult result = pointService.usePoints(customerId, 2000);
 
-        // then
         assertThat(result.isSuccess()).isTrue();
         assertThat(result.getUsedPoints()).isEqualTo(2000);
         assertThat(result.getRemainingPoints()).isEqualTo(3000);
@@ -56,7 +52,6 @@ class PointServiceTest {
     @DisplayName("포인트 부족 시 사용 실패")
     @Test
     void givenInsufficientPoints_whenUsePoints_thenReturnFailure() {
-        // given
         UUID customerId = UUID.randomUUID();
         Customer customer = Customer.builder()
                 .id(customerId)
@@ -66,10 +61,8 @@ class PointServiceTest {
         given(customerRepository.findById(customerId))
                 .willReturn(Optional.of(customer));
 
-        // when
         PointResult result = pointService.usePoints(customerId, 2000);
 
-        // then
         assertThat(result.isSuccess()).isFalse();
         assertThat(result.getErrorMessage()).contains("포인트가 부족합니다");
         assertThat(customer.getPoints()).isEqualTo(1000); // 원래 포인트 유지
@@ -79,16 +72,13 @@ class PointServiceTest {
     @DisplayName("존재하지 않는 고객 포인트 사용 실패")
     @Test
     void givenNonExistentCustomer_whenUsePoints_thenReturnFailure() {
-        // given
         UUID nonExistentCustomerId = UUID.randomUUID();
 
         given(customerRepository.findById(nonExistentCustomerId))
                 .willReturn(Optional.empty());
 
-        // when
         PointResult result = pointService.usePoints(nonExistentCustomerId, 1000);
 
-        // then
         assertThat(result.isSuccess()).isFalse();
         assertThat(result.getErrorMessage()).contains("고객을 찾을 수 없습니다");
     }
@@ -96,13 +86,10 @@ class PointServiceTest {
     @DisplayName("0 포인트 사용 시 실패")
     @Test
     void givenZeroPoints_whenUsePoints_thenReturnFailure() {
-        // given
         UUID customerId = UUID.randomUUID();
 
-        // when
         PointResult result = pointService.usePoints(customerId, 0);
 
-        // then
         assertThat(result.isSuccess()).isFalse();
         assertThat(result.getErrorMessage()).contains("유효하지 않은 포인트 사용 요청입니다");
         then(customerRepository).should(never()).findById(any(UUID.class));
@@ -112,13 +99,10 @@ class PointServiceTest {
     @DisplayName("음수 포인트 사용 시 실패")
     @Test
     void givenNegativePoints_whenUsePoints_thenReturnFailure() {
-        // given
         UUID customerId = UUID.randomUUID();
 
-        // when
         PointResult result = pointService.usePoints(customerId, -1000);
 
-        // then
         assertThat(result.isSuccess()).isFalse();
         assertThat(result.getErrorMessage()).contains("유효하지 않은 포인트 사용 요청입니다");
         then(customerRepository).should(never()).findById(any(UUID.class));
@@ -128,13 +112,10 @@ class PointServiceTest {
     @DisplayName("null 포인트 사용 시 실패")
     @Test
     void givenNullPoints_whenUsePoints_thenReturnFailure() {
-        // given
         UUID customerId = UUID.randomUUID();
 
-        // when
         PointResult result = pointService.usePoints(customerId, null);
 
-        // then
         assertThat(result.isSuccess()).isFalse();
         assertThat(result.getErrorMessage()).contains("유효하지 않은 포인트 사용 요청입니다");
         then(customerRepository).should(never()).findById(any(UUID.class));
@@ -144,13 +125,10 @@ class PointServiceTest {
     @DisplayName("null 고객 ID로 포인트 사용 시 실패")
     @Test
     void givenNullCustomerId_whenUsePoints_thenReturnFailure() {
-        // given
         Integer pointsToUse = 1000;
 
-        // when
         PointResult result = pointService.usePoints(null, pointsToUse);
 
-        // then
         assertThat(result.isSuccess()).isFalse();
         assertThat(result.getErrorMessage()).contains("유효하지 않은 포인트 사용 요청입니다");
         then(customerRepository).should(never()).findById(any(UUID.class));
@@ -160,7 +138,6 @@ class PointServiceTest {
     @DisplayName("포인트 적립 성공")
     @Test
     void givenValidCustomer_whenRefundPoints_thenReturnSuccess() {
-        // given
         UUID customerId = UUID.randomUUID();
         Customer customer = Customer.builder()
                 .id(customerId)
@@ -172,10 +149,8 @@ class PointServiceTest {
         given(customerRepository.save(any(Customer.class)))
                 .willReturn(customer);
 
-        // when
         PointResult result = pointService.refundPoints(customerId, 500);
 
-        // then
         assertThat(result.isSuccess()).isTrue();
         assertThat(result.getUsedPoints()).isEqualTo(500);
         assertThat(result.getRemainingPoints()).isEqualTo(1500);
@@ -186,7 +161,6 @@ class PointServiceTest {
     @DisplayName("포인트 사용 가능 여부 확인 - 충분한 포인트")
     @Test
     void givenSufficientPoints_whenCanUsePoints_thenReturnTrue() {
-        // given
         UUID customerId = UUID.randomUUID();
         Customer customer = Customer.builder()
                 .id(customerId)
@@ -196,17 +170,14 @@ class PointServiceTest {
         given(customerRepository.findById(customerId))
                 .willReturn(Optional.of(customer));
 
-        // when
         boolean canUse = pointService.canUsePoints(customerId, 2000);
 
-        // then
         assertThat(canUse).isTrue();
     }
 
     @DisplayName("포인트 사용 가능 여부 확인 - 부족한 포인트")
     @Test
     void givenInsufficientPoints_whenCanUsePoints_thenReturnFalse() {
-        // given
         UUID customerId = UUID.randomUUID();
         Customer customer = Customer.builder()
                 .id(customerId)
@@ -216,17 +187,14 @@ class PointServiceTest {
         given(customerRepository.findById(customerId))
                 .willReturn(Optional.of(customer));
 
-        // when
         boolean canUse = pointService.canUsePoints(customerId, 2000);
 
-        // then
         assertThat(canUse).isFalse();
     }
 
     @DisplayName("null 값으로 포인트 사용 가능 여부 확인 시 false 반환")
     @Test
     void givenNullValues_whenCanUsePoints_thenReturnFalse() {
-        // when & then
         assertThat(pointService.canUsePoints(null, 1000)).isFalse();
         assertThat(pointService.canUsePoints(UUID.randomUUID(), null)).isFalse();
         assertThat(pointService.canUsePoints(null, null)).isFalse();
